refactor: extract data statistics table parsing into separate file
